refactor(ui): extract <Conversation> container + transcript helpers
