+++ ui/components/chats.jsx
@@ -517,550 +517,44 @@ function CT_AgentSwitcher({ chatId, currentAgentId, pushToast, placement = "down
 // ============================================================================
 
 function ChatDetail({ chatId, onBack, pushToast }) {
-  const { useResource, useMutation, useViewport, apiFetch } = window.primerApi;
+  const { useResource, useViewport, apiFetch } = window.primerApi;
   const { isMobile } = useViewport();
   const cid = chatId;
   // Mobile-only kebab actions sheet (collapses Token meter + Compact +
   // status pill that don't fit in the slim mobile header).
   const [actionsOpen, setActionsOpen] = React.useState(false);
 
-  const [messages, setMessages] = React.useState([]);
-  const [lastSeq, setLastSeq] = React.useState(0);
-  // Oldest seq currently loaded; null until the initial tail fetch
-  // completes. Drives the scroll-up lazy-loader's `before_seq` cursor.
-  const [oldestSeq, setOldestSeq] = React.useState(null);
-  // false once the tail fetch (or a later older-page fetch) returns
-  // fewer rows than asked for — we've hit the top of history.
-  const [hasMoreOlder, setHasMoreOlder] = React.useState(false);
-  const [loadingOlder, setLoadingOlder] = React.useState(false);
-  // Set by the initial-load effect when the first batch is in the
-  // store; gates the WebSocket open so the WS cursor can be set to
-  // the highest already-loaded seq and skip a redundant full replay.
-  const [initialLoadedSeq, setInitialLoadedSeq] = React.useState(null);
-  const [wsState, setWsState] = React.useState("connecting");
-  const [composer, setComposer] = React.useState("");
-  const [historyError, setHistoryError] = React.useState(null);
-  // Live token-usage snapshot — driven by `"usage"` WS envelopes the
-  // worker emits after each turn. Drives the header TokenMeter pill;
-  // values stay at 0 until the first envelope lands so the meter
-  // renders dimmed but present.
-  const [usage, setUsage] = React.useState({ input_tokens: 0, output_tokens: 0, context_length: 0 });
-  const [compactInFlight, setCompactInFlight] = React.useState(false);
-  // Set true the moment the user sends a frame; cleared when any
-  // assistant_token / tool_call / done / error row arrives. Drives
-  // the "Thinking..." placeholder so the operator sees the system
-  // is responding before the first delta lands.
-  const [waitingForReply, setWaitingForReply] = React.useState(false);
-  // Pending file attachments. Each entry: {id, file, name, mime, kind, dataB64, preview}.
-  const [attachments, setAttachments] = React.useState([]);
-  const wsRef = React.useRef(null);
-  const scrollRef = React.useRef(null);
-  const fileInputRef = React.useRef(null);
+  // Status bag reported by <Conversation> (Task B2 moved the WS/data
+  // lifecycle wholesale into that component) — this host only needs
+  // enough of it to drive its own page chrome: the TokenMeter, the
+  // connection badge, the status pill, the mobile kebab sheet, and
+  // the "chat not found" gate below.
+  const [convStatus, setConvStatus] = React.useState({
+    wsState: "connecting",
+    usage: { input_tokens: 0, output_tokens: 0, context_length: 0 },
+    compactInFlight: false,
+    historyError: null,
+    requestCompact: null,
+  });
 
   // Chat row (status, agent_id) — small one-shot fetch + light polling
-  // so the header pill mirrors a server-driven "ended" state.
+  // so the header pill mirrors a server-driven "ended" state. Shares
+  // the useResource cache entry with <Conversation>'s own fetch of
+  // the same key (see foundation/use-resource.js), so this does NOT
+  // double the network traffic.
   const chat = useResource(
     `chat-detail:${cid}`,
     (s) => apiFetch("GET", `/chats/${encodeURIComponent(cid)}`, null, { signal: s }),
     { pollMs: 10000, deps: [cid] }
   );
 
-  // Initial REST load — fetches the TAIL of the history so the
-  // renderer can scroll straight to the bottom without dragging
-  // through thousands of older rows. The WS then opens with
-  // cursor=<lastSeq> below so it streams only new messages, not a
-  // redundant full replay. Older rows lazy-load on scroll-up.
-  //
-  // Page size is 200 (the server pagination cap) and we keep paging
-  // BACKWARDS until either:
-  //   - we've crossed at least one `user_message` (so the operator
-  //     sees the prior turn, not just the tail of one long response —
-  //     each LLM token streams as its own assistant_token row, so a
-  //     single reply can span 100+ rows and a naive single-page
-  //     tail-load returns only fragments of that one message); OR
-  //   - we've exhausted the chat; OR
-  //   - we've hit the safety cap of TAIL_MAX_PAGES iterations.
-  const TAIL_PAGE_SIZE = 200;
-  const TAIL_MAX_PAGES = 6; // ~1200 rows max on initial load
-  // Pagination ceiling is 2**53 - 1; the server caps int Query at the
-  // 64-bit boundary, but Number.MAX_SAFE_INTEGER is unambiguous and
-  // matches every persisted seq.
-  const SENTINEL_TAIL_SEQ = Number.MAX_SAFE_INTEGER;
-  React.useEffect(() => {
-    let cancelled = false;
-    setMessages([]);
-    setLastSeq(0);
-    setOldestSeq(null);
-    setHasMoreOlder(false);
-    setInitialLoadedSeq(null);
-    (async () => {
-      try {
-        let collected = []; // prepended each iteration so it stays ASC
-        let cursor = SENTINEL_TAIL_SEQ;
-        let hasMore = true;
-        let foundUserMsg = false;
-        for (let i = 0; i < TAIL_MAX_PAGES; i++) {
-          const data = await apiFetch(
-            "GET",
-            `/chats/${encodeURIComponent(cid)}/messages?before_seq=${cursor}&limit=${TAIL_PAGE_SIZE}`,
-          );
-          if (cancelled) return;
-          const items = (data && data.items) || [];
-          if (items.length === 0) {
-            hasMore = false;
-            break;
-          }
-          // REST returns ChatMessage rows with kind-specific fields
-          // nested under `payload`. WS frames spread payload into the
-          // top-level (see chats router _message_to_wire). Flatten on
-          // load to keep both sources homogeneous.
-          const flat = items.map((row) => {
-            const payload = row.payload && typeof row.payload === "object" ? row.payload : {};
-            return { ...payload, ...row };
-          });
-          collected = [...flat, ...collected];
-          foundUserMsg = foundUserMsg || flat.some((r) => r.kind === "user_message");
-          if (items.length < TAIL_PAGE_SIZE) {
-            hasMore = false;
-            break;
-          }
-          if (foundUserMsg) {
-            break; // We've reached at least one full turn — operator can see context.
-          }
-          cursor = flat[0].seq || cursor;
-        }
-        setMessages(collected);
-        if (collected.length > 0) {
-          const last = collected[collected.length - 1].seq || 0;
-          const first = collected[0].seq || 0;
-          setLastSeq(last);
-          setOldestSeq(first);
-          setHasMoreOlder(hasMore);
-          setInitialLoadedSeq(last);
-        } else {
-          // Empty chat — open WS with cursor 0 to catch the very
-          // first message that may land between mount and any send.
-          setInitialLoadedSeq(0);
-        }
-      } catch (err) {
-        if (cancelled) return;
-        setHistoryError(err);
-        // Fall through to opening the WS at cursor 0 so the user can
-        // still send and see new messages even if history failed.
-        setInitialLoadedSeq(0);
-      }
-    })();
-    return () => { cancelled = true; };
-  }, [cid]); // eslint-disable-line react-hooks/exhaustive-deps
-
-  // Lazy-load older messages on scroll-up. Captures scroll geometry
-  // before the prepend; a useLayoutEffect (below) restores the
-  // visible position synchronously between React's commit and the
-  // browser's paint, so the user never sees the intermediate
-  // (wrong-scroll) frame.
-  const loadingOlderRef = React.useRef(false);
-  // When non-null, the next layout effect should snap scrollTop so
-  // the previously-visible content stays put. Shape:
-  // { oldScrollHeight: number, oldScrollTop: number }.
-  const pendingPrependRef = React.useRef(null);
-  const loadOlder = React.useCallback(async () => {
-    if (loadingOlderRef.current) return;
-    if (!hasMoreOlder || oldestSeq == null || oldestSeq <= 1) return;
-    loadingOlderRef.current = true;
-    setLoadingOlder(true);
-    try {
-      const data = await apiFetch(
-        "GET",
-        `/chats/${encodeURIComponent(cid)}/messages?before_seq=${oldestSeq}&limit=${TAIL_PAGE_SIZE}`,
-      );
-      const items = (data && data.items) || [];
-      if (items.length === 0) {
-        setHasMoreOlder(false);
-        return;
-      }
-      const flat = items.map((row) => {
-        const payload = row.payload && typeof row.payload === "object" ? row.payload : {};
-        return { ...payload, ...row };
-      });
-      // Capture scroll geometry RIGHT before triggering the prepend
-      // re-render. The layout effect (keyed on messages.length) reads
-      // this ref synchronously after React commits and adjusts
-      // scrollTop before the browser paints — no visible jump.
-      const el = scrollRef.current;
-      if (el) {
-        pendingPrependRef.current = {
-          oldScrollHeight: el.scrollHeight,
-          oldScrollTop: el.scrollTop,
-        };
-      }
-      setMessages((prev) => [...flat, ...prev]);
-      setOldestSeq(flat[0].seq || oldestSeq);
-      setHasMoreOlder(items.length === TAIL_PAGE_SIZE);
-    } catch (err) {
-      if (typeof pushToast === "function") {
-        pushToast({ kind: "error", title: "Load older failed", detail: err?.message || "" });
-      }
-    } finally {
-      setLoadingOlder(false);
-      loadingOlderRef.current = false;
-    }
-  }, [cid, oldestSeq, hasMoreOlder, apiFetch, pushToast]);
-
-  // Synchronous scroll-position restoration after a prepend. Runs
-  // post-commit, pre-paint — so the browser paints exactly once with
-  // the correct scrollTop, and the previously-visible content stays
-  // pinned to the same on-screen pixel row.
-  React.useLayoutEffect(() => {
-    const pending = pendingPrependRef.current;
-    if (pending == null) return;
-    pendingPrependRef.current = null;
-    const el = scrollRef.current;
-    if (!el) return;
-    const delta = el.scrollHeight - pending.oldScrollHeight;
-    el.scrollTop = pending.oldScrollTop + delta;
-  }, [messages.length]);
-
-  // WS lifecycle — opens once the initial REST tail-load has settled,
-  // using the tail's highest seq as the cursor so the server only
-  // streams NEW frames (no redundant full-history replay). The init
-  // gate (`initialLoadedSeq != null`) tolerates an empty-chat or
-  // failed-load fallback to cursor 0.
-  //
-  // Reconnect: an unexpected close (network blip, server restart)
-  // triggers exponential-backoff reconnect (1s -> 2s -> 4s ... 30s cap).
-  // Reconnection reuses the last received seq as the cursor so no
-  // frames are missed or duplicated. Terminal close codes (4404, 4410)
-  // do not reconnect.
-  React.useEffect(() => {
-    if (initialLoadedSeq == null) return;
-    let intentional = false;
-    let backoffMs = 1000;
-    const MAX_BACKOFF_MS = 30000;
-    let reconnectTimer = null;
-    // Track the highest seq received in this effect's lifetime.
-    // Starts at initialLoadedSeq so the first connection opens with the
-    // correct cursor; updated by onmessage so reconnects resume from the
-    // last received frame.
-    let latestSeq = initialLoadedSeq;
-
-    function connect() {
-      if (intentional) return;
-      const proto = window.location.protocol === "https:" ? "wss:" : "ws:";
-      const url = `${proto}//${window.location.host}/v1/chats/${encodeURIComponent(cid)}/ws?cursor=${latestSeq}`;
-      let ws;
-      try {
-        ws = new WebSocket(url);
-      } catch (e) {
-        setWsState("closed");
-        return;
-      }
-      wsRef.current = ws;
-      setWsState("connecting");
-
-      ws.onopen = () => {
-        setWsState("open");
-        backoffMs = 1000; // reset on successful connect
-      };
-
-      ws.onmessage = (ev) => {
-        let msg;
-        try { msg = JSON.parse(ev.data); } catch { return; }
-        if (!msg || typeof msg !== "object") return;
-
-        if (msg.kind === "error" && typeof msg.seq !== "number") {
-          // Protocol-level error frame (not a persisted row).
-          if (typeof pushToast === "function") {
-            pushToast({
-              kind: "error",
-              title: msg.code || "WebSocket error",
-              detail: msg.message || "",
-            });
-          }
-          return;
-        }
-        if (msg.kind === "pong") return;
-
-        // Token usage envelope (no seq). Emitted by the worker after each
-        // assistant turn. Snapshot drives the header TokenMeter pill.
-        if (msg.kind === "usage" && typeof msg.seq !== "number") {
-          setUsage({
-            input_tokens: Number(msg.input_tokens) || 0,
-            output_tokens: Number(msg.output_tokens) || 0,
-            context_length: Number(msg.context_length) || 0,
-          });
-          return;
-        }
-
-        // Compaction envelope (no seq). Server tells us a compaction
-        // pass just happened; surface in three places so the operator
-        // sees it: an in-stream marker row, a success toast, and an
-        // immediate TokenMeter update reflecting the new prompt size.
-        //
-        // Field names mirror the server envelope shape
-        // (tokens_before, tokens_after) per
-        // primer/api/routers/chats.py::_compaction_envelope.
-        //
-        // tokens_after IS the post-compaction context size - per
-        // primer/agent/compaction.py::_full_compact + _estimate_tokens,
-        // it is computed over the FULL new history (summary + retained
-        // tail), not just the summary payload. Pinning the meter to it
-        // is correct: that's the prompt the next assistant turn carries.
-        // Compaction envelope. The server translates the persisted
-        // compaction_marker row into this 'compaction' envelope and sends
-        // it WITH the row's seq (primer/api/routers/chats.py::
-        // _compaction_envelope). Handle it regardless of whether a seq is
-        // present so the marker, the TokenMeter update, and the toast all
-        // fire. Earlier code required a missing seq and so silently
-        // dropped every server-sent compaction (the meter never moved and
-        // no completion marker appeared).
-        if (msg.kind === "compaction") {
-          const beforeT = Number(msg.tokens_before) || 0;
-          const afterT = Number(msg.tokens_after) || 0;
-          const markerSeq = typeof msg.seq === "number"
-            ? msg.seq
-            : `compaction-${Date.now()}`;
-          // Clear the in-progress flag and append the completion marker
-          // (de-duped by seq against the cursor replay).
-          setCompactInFlight(false);
-          setMessages((prev) => prev.some((m) => m.seq === markerSeq)
-            ? prev
-            : [...prev, {
-                kind: "compaction_marker",
-                seq: markerSeq,
-                tokens_before: beforeT,
-                tokens_after: afterT,
-                reason: msg.reason || "",
-              }]);
-          // Update the context meter to the post-compaction prompt size so
-          // the top-right indicator reflects the smaller window immediately.
-          if (afterT > 0) {
-            setUsage((prev) => ({
-              ...prev,
-              input_tokens: afterT,
-            }));
-          }
-          if (typeof pushToast === "function") {
-            const saved = beforeT > 0 ? Math.max(0, beforeT - afterT) : 0;
-            pushToast({
-              kind: "success",
-              title: "Compaction complete",
-              detail: beforeT > 0
-                ? `${beforeT.toLocaleString()} -> ${afterT.toLocaleString()} tokens`
-                  + (saved > 0 ? ` (saved ${saved.toLocaleString()})` : "")
-                : null,
-            });
-          }
-          return;
-        }
-
-        // Any persisted row carries seq → append + advance cursor. We
-        // de-dupe against the initial REST replay because both sources
-        // may overlap (REST loaded seqs 1..N, WS replays seq>0).
-        if (typeof msg.seq === "number") {
-          if (msg.seq > latestSeq) latestSeq = msg.seq;
-          setMessages((prev) => {
-            if (prev.some((p) => p.seq === msg.seq)) return prev;
-            return [...prev, msg];
-          });
-          setLastSeq((prev) => (msg.seq > prev ? msg.seq : prev));
-          // The agent is now producing output (or finished); drop the
-          // thinking placeholder. user_message echoes from a previous
-          // turn don't reach here because we only get rows after the
-          // server processes our outbound frame.
-          if (msg.kind !== "user_message") {
-            setWaitingForReply(false);
-          }
-        }
-      };
-
-      ws.onclose = (ev) => {
-        wsRef.current = null;
-        setWsState("closed");
-        // Terminal codes: do not reconnect.
-        if (ev.code === 4404) {
-          if (typeof pushToast === "function") {
-            pushToast({ kind: "error", title: "Chat not found", detail: ev.reason || cid });
-          }
-          return;
-        }
-        if (ev.code === 4410) {
-          if (typeof pushToast === "function") {
-            pushToast({ kind: "warning", title: "Chat ended", detail: ev.reason || cid });
-          }
-          return;
-        }
-        // Unexpected close - reconnect with exponential backoff.
-        if (!intentional) {
-          reconnectTimer = setTimeout(() => {
-            backoffMs = Math.min(backoffMs * 2, MAX_BACKOFF_MS);
-            connect();
-          }, backoffMs);
-        }
-      };
-
-      ws.onerror = () => {
-        // Browsers report a generic ErrorEvent and then close - onclose
-        // handles user-facing messaging and reconnect scheduling.
-      };
-    }
-
-    connect();
-
-    return () => {
-      intentional = true;
-      if (reconnectTimer != null) clearTimeout(reconnectTimer);
-      try { wsRef.current && wsRef.current.close(); } catch { /* no-op */ }
-      wsRef.current = null;
-    };
-  }, [cid, initialLoadedSeq]); // eslint-disable-line react-hooks/exhaustive-deps
-
-  // Auto-scroll to bottom only when the tail grows (initial load or
-  // a live frame). `lastSeq` is monotone-increasing on appends, so
-  // depending on it instead of `messages` means scroll-up prepends
-  // (which keep lastSeq unchanged) don't yank the user back down.
-  const stickToBottomRef = React.useRef(true);
-  const onScroll = React.useCallback(() => {
-    const el = scrollRef.current;
-    if (!el) return;
-    const distance = el.scrollHeight - el.scrollTop - el.clientHeight;
-    stickToBottomRef.current = distance < 80;
-    // Near the top → fetch the next older page. The 100px threshold
-    // gives a small buffer so the user sees the loading indicator
-    // before they bottom-out the scroll.
-    if (el.scrollTop < 100) {
-      loadOlder();
-    }
-  }, [loadOlder]);
-  React.useEffect(() => {
-    if (!scrollRef.current || !stickToBottomRef.current) return;
-    const el = scrollRef.current;
-    const raf = requestAnimationFrame(() => {
-      el.scrollTop = el.scrollHeight;
-    });
-    return () => cancelAnimationFrame(raf);
-  }, [lastSeq, waitingForReply]);
-
-  // Returns true if the frame was enqueued; false if the socket was not
-  // ready (user-facing toast fires on false so the text is preserved).
-  const sendMessage = (text, atts) => {
-    const ws = wsRef.current;
-    if (!ws || ws.readyState !== 1) {
-      if (typeof pushToast === "function") {
-        pushToast({ kind: "error", title: "Not connected", detail: "WebSocket is not open" });
-      }
-      return false;
-    }
-    const frame = { kind: "user_message" };
-    if (text) frame.content = text;
-    if (atts && atts.length > 0) {
-      frame.parts = atts.map((a) => ({
-        type: a.kind,           // "image" or "document"
-        data: a.dataB64,
-        mime_type: a.mime,
-        ...(a.kind === "document" && a.name ? { filename: a.name } : {}),
-      }));
-    }
-    ws.send(JSON.stringify(frame));
-    setWaitingForReply(true);
-    return true;
-  };
-
-  // Clear the composer and attachments only after a successful send so a
-  // failed send (WS not open) leaves the user's text intact.
-  const onSubmitComposer = () => {
-    const text = composer.trim();
-    if (!text && attachments.length === 0) return;
-    const sent = sendMessage(text, attachments);
-    if (sent) {
-      setComposer("");
-      setAttachments([]);
-    }
-  };
-
-  // Operator-triggered compaction. POSTs to the chat's compact
-  // endpoint; the server runs the configured compaction prompt
-  // against the conversation and emits a `"compaction"` envelope on
-  // the WS once finished — that envelope flows through onmessage
-  // above and appends a `compaction_marker` row to the timeline.
-  const handleCompact = async () => {
-    if (compactInFlight) return;
-    setCompactInFlight(true);
-    try {
-      await apiFetch("POST", `/chats/${encodeURIComponent(cid)}/compact`);
-      if (typeof pushToast === "function") {
-        pushToast({ kind: "success", title: "Compaction started" });
-      }
-    } catch (err) {
-      if (typeof pushToast === "function") {
-        pushToast({
-          kind: "error",
-          title: err?.title || "Compact failed",
-          detail: err?.detail || err?.message,
-          requestId: err?.requestId,
-        });
-      }
-    } finally {
-      setCompactInFlight(false);
-    }
-  };
-
-  // ---- File picking + base64 encoding ----------------------------------
-  // The chat protocol's user_message.parts list expects each binary part
-  // to carry `data` (base64), `mime_type`, and (for documents) a
-  // filename hint. Browsers read files as a Blob; FileReader gives us
-  // a data URL we can split into the b64 tail.
-  const MAX_ATTACHMENT_BYTES = 8 * 1024 * 1024; // 8 MiB — keeps the WS frame sane.
-  const handleFilesPicked = async (fileList) => {
-    if (!fileList || fileList.length === 0) return;
-    const next = [...attachments];
-    for (const file of Array.from(fileList)) {
-      if (file.size > MAX_ATTACHMENT_BYTES) {
-        if (typeof pushToast === "function") {
-          pushToast({
-            kind: "error",
-            title: "File too large",
-            detail: `${file.name} is ${(file.size / 1024 / 1024).toFixed(1)} MiB; limit is 8 MiB.`,
-          });
-        }
-        continue;
-      }
-      const mime = file.type || "application/octet-stream";
-      const isImage = mime.startsWith("image/");
-      const dataB64 = await new Promise((resolve, reject) => {
-        const reader = new FileReader();
-        reader.onload = () => {
-          const result = String(reader.result || "");
-          const comma = result.indexOf(",");
-          resolve(comma >= 0 ? result.slice(comma + 1) : result);
-        };
-        reader.onerror = () => reject(reader.error);
-        reader.readAsDataURL(file);
-      });
-      next.push({
-        id: `att-${Date.now()}-${Math.random().toString(36).slice(2, 8)}`,
-        name: file.name,
-        mime,
-        kind: isImage ? "image" : "document",
-        dataB64,
-        preview: isImage ? `data:${mime};base64,${dataB64}` : null,
-        size: file.size,
-      });
-    }
-    setAttachments(next);
-    if (fileInputRef.current) fileInputRef.current.value = "";
-  };
-
-  const removeAttachment = (id) => {
-    setAttachments((prev) => prev.filter((a) => a.id !== id));
-  };
-
-  const wsBadge = wsState === "open"
+  const wsBadge = convStatus.wsState === "open"
     ? <span className="pill pill-running" title="WebSocket open"><span className="dot"></span>live</span>
-    : wsState === "connecting"
+    : convStatus.wsState === "connecting"
       ? <span className="pill pill-paused" title="WebSocket connecting"><span className="dot"></span>connecting</span>
       : <span className="pill pill-ended" title="WebSocket closed"><span className="dot"></span>offline</span>;
 
-  if (historyError && historyError.status === 404) {
+  if (convStatus.historyError && convStatus.historyError.status === 404) {
     return (
       <Banner
         kind="error"
@@ -1086,8 +580,8 @@ function ChatDetail({ chatId, onBack, pushToast }) {
         // 100vh on iOS/Android Chrome is the *large* viewport that
         // ignores the address bar, which makes the inner scroll
         // container overflow the visible area and the BODY scrolls
-        // instead of `scrollRef.current` — breaking pull-up-to-load-
-        // older and lazy-prepend.
+        // instead of the transcript's own scroll container — breaking
+        // pull-up-to-load-older and lazy-prepend.
         //
         // Mobile deduction is just the global topbar (48px) plus a
         // few px of breathing room; the chat panel renders its own
@@ -1136,14 +630,14 @@ function ChatDetail({ chatId, onBack, pushToast }) {
           )}
           <div className="right" style={{ display: "flex", gap: 8, alignItems: "center" }}>
             <window.TokenMeter
-              inputTokens={usage.input_tokens}
-              contextLength={usage.context_length}
-              onCompact={chatStatus === "ended" ? null : handleCompact}
-              compactDisabled={compactInFlight || wsState !== "open"}
+              inputTokens={convStatus.usage.input_tokens}
+              contextLength={convStatus.usage.context_length}
+              onCompact={chatStatus === "ended" ? null : convStatus.requestCompact}
+              compactDisabled={convStatus.compactInFlight || convStatus.wsState !== "open"}
               compactTooltip={
-                compactInFlight
+                convStatus.compactInFlight
                   ? "Compaction in progress…"
-                  : wsState !== "open"
+                  : convStatus.wsState !== "open"
                     ? "WebSocket offline"
                     : ""
               }
@@ -1155,173 +649,14 @@ function ChatDetail({ chatId, onBack, pushToast }) {
           </div>
         </div>
         )}
-        <div ref={scrollRef} onScroll={onScroll} style={{ flex: 1, overflow: "auto", padding: "18px 24px", minHeight: 0, minWidth: 0 }}>
-          {(loadingOlder || hasMoreOlder) && messages.length > 0 && (
-            <div
-              className="muted text-sm"
-              style={{ textAlign: "center", padding: "6px 0 12px", fontSize: 11 }}
-              data-testid="chat-load-older"
-            >
-              {loadingOlder ? "Loading older…" : "Scroll up to load older"}
-            </div>
-          )}
-          {messages.length === 0 && !historyError && (
-            <div className="muted text-sm" style={{ textAlign: "center", padding: 24 }}>
-              {wsState === "connecting" ? "Connecting…" : "No messages yet. Say hello to the agent."}
-            </div>
-          )}
-          {CT_coalesceMessages(messages).map((m) =>
-            m.kind === "assistant_message" ? (
-              <Message key={`am-${m.startSeq}-${m.endSeq}`} m={m} />
-            ) : (
-              <Message key={`${m.seq}-${m.kind}`} m={m} />
-            )
-          )}
-
-          {/* Thinking indicator — local flag for the freshly-sent
-              turn, plus a turn-status fallback driven by the chat row.
-              Visible whenever the turn is in flight AND the last
-              persisted row is NOT itself the agent's active response
-              or a terminal close-out. This covers the gap between a
-              tool_call landing and the next assistant_token streaming
-              — without it the operator sees the indicator vanish at
-              the first tool call even though the worker is still busy
-              processing the result. */}
-          {(() => {
-            const lastRow = messages.length > 0 ? messages[messages.length - 1] : null;
-            const turnInFlight = chatRow
-              && (chatRow.turn_status === "claimable" || chatRow.turn_status === "running");
-            // Rows that mean "agent is currently producing visible
-            // output, no thinking placeholder needed" or "turn closed
-            // out, definitely no placeholder."
-            const QUIET_LAST_KINDS = new Set([
-              "assistant_token",  // currently streaming a response
-              "done",
-              "error",
-              "cancelled",
-              "yielded",
-            ]);
-            const lastIsQuiet = lastRow && QUIET_LAST_KINDS.has(lastRow.kind);
-            const showThinking =
-              waitingForReply
-              || (turnInFlight && !lastIsQuiet);
-            return showThinking ? <CT_ThinkingBubble /> : null;
-          })()}
-
-          {/* Compaction in-progress indicator. Shown from the moment the
-              operator clicks Compact until the server's compaction
-              envelope lands (which clears compactInFlight and appends the
-              completion marker). Without it the only feedback was the
-              disabled button in the header. */}
-          {compactInFlight && (
-            <div style={{
-              display: "flex", alignItems: "center", gap: 8,
-              margin: "12px auto", padding: "6px 14px",
-              border: "1px dashed var(--border)", borderRadius: 14,
-              width: "fit-content", fontSize: 12,
-              color: "var(--text-3)", background: "var(--bg-1, var(--bg))",
-            }}>
-              <Icon name="compress" size={12} className="muted" />
-              <span>Compacting conversation history…</span>
-            </div>
-          )}
-        </div>
-
-        {/* Pending-attachments strip — visible only when the composer
-            has files queued. Each chip carries an image thumbnail or a
-            document icon + filename + size; clicking ×  drops it. */}
-        {attachments.length > 0 && (
-          <div
-            style={{
-              borderTop: "1px solid var(--border)",
-              padding: "10px 14px",
-              display: "flex",
-              flexWrap: "wrap",
-              gap: 8,
-            }}
-          >
-            {attachments.map((a) => (
-              <CT_AttachmentChip
-                key={a.id}
-                attachment={a}
-                onRemove={() => removeAttachment(a.id)}
-              />
-            ))}
-          </div>
-        )}
-
-        <div
-          className={isMobile ? "composer-sticky" : ""}
-          style={
-            isMobile
-              ? {
-                  display: "flex",
-                  gap: 8,
-                  alignItems: "stretch",
-                }
-              : {
-                  borderTop: "1px solid var(--border)",
-                  padding: 14,
-                  display: "flex",
-                  gap: 8,
-                  alignItems: "stretch",
-                }
-          }
-        >
-          <CT_AgentSwitcher
-            chatId={cid}
-            currentAgentId={chatAgent}
-            pushToast={pushToast}
-            placement="up"
-            disabled={chatStatus === "ended"}
-            triggerStyle={{ padding: "0 12px", borderRadius: 6, alignSelf: "stretch" }}
-          />
-          <button
-            type="button"
-            title="Attach files (images, PDFs)"
-            data-testid="chat-attach-btn"
-            onClick={() => fileInputRef.current && fileInputRef.current.click()}
-            disabled={chatStatus === "ended"}
-            style={{
-              background: "transparent",
-              border: "1px solid var(--border)",
-              borderRadius: 6,
-              padding: "0 10px",
-              color: "var(--text-2)",
-              cursor: chatStatus === "ended" ? "not-allowed" : "pointer",
-              display: "flex",
-              alignItems: "center",
-              opacity: chatStatus === "ended" ? 0.5 : 1,
-            }}
-          >
-            <Icon name="paperclip" size={14} />
-          </button>
-          <input
-            ref={fileInputRef}
-            type="file"
-            multiple
-            accept="image/*,application/pdf"
-            style={{ display: "none" }}
-            onChange={(e) => handleFilesPicked(e.target.files)}
-          />
-          <textarea
-            className="textarea"
-            value={composer}
-            onChange={(e) => setComposer(e.target.value)}
-            placeholder={chatStatus === "ended" ? "This chat has ended." : "Send a message…"}
-            rows={2}
-            style={{ flex: 1, resize: "none" }}
-            disabled={chatStatus === "ended"}
-            onKeyDown={(e) => { if (e.key === "Enter" && !e.shiftKey) { e.preventDefault(); onSubmitComposer(); } }}
-          />
-          <Btn
-            kind="primary"
-            icon="send"
-            disabled={(!composer.trim() && attachments.length === 0) || chatStatus === "ended" || wsState !== "open"}
-            onClick={onSubmitComposer}
-            style={{ alignSelf: "stretch", paddingLeft: 16, paddingRight: 16 }}
-          >Send</Btn>
-        </div>
+        <Conversation
+          chatId={cid}
+          pushToast={pushToast}
+          onStatus={setConvStatus}
+          headerSlot={null}
+          rightChromeSlot={null}
+          showSchemaPanel={false}
+        />
       </div>
 
       {/* Mobile-only kebab actions sheet — collapses TokenMeter,
@@ -1335,17 +670,17 @@ function ChatDetail({ chatId, onBack, pushToast }) {
         >
           <div className="col" style={{ gap: 12 }}>
             <window.TokenMeter
-              inputTokens={usage.input_tokens}
-              contextLength={usage.context_length}
+              inputTokens={convStatus.usage.input_tokens}
+              contextLength={convStatus.usage.context_length}
               onCompact={chatStatus === "ended" ? null : () => {
                 setActionsOpen(false);
-                handleCompact();
+                convStatus.requestCompact && convStatus.requestCompact();
               }}
-              compactDisabled={compactInFlight || wsState !== "open"}
+              compactDisabled={convStatus.compactInFlight || convStatus.wsState !== "open"}
               compactTooltip={
-                compactInFlight
+                convStatus.compactInFlight
                   ? "Compaction in progress…"
-                  : wsState !== "open"
+                  : convStatus.wsState !== "open"
                     ? "WebSocket offline"
                     : ""
               }
@@ -1366,53 +701,13 @@ function ChatDetail({ chatId, onBack, pushToast }) {
 }
 
 // ============================================================================
-// Helpers: assistant-token coalescing + thinking indicator + attachment chip
+// Helpers: thinking indicator + attachment chip
 // ============================================================================
-
-// Walk `messages` in order and merge any run of consecutive
-// `assistant_token` rows into one synthetic "assistant_message"
-// entry whose `text` is the concatenation of the run's deltas. Any
-// other row passes through unchanged. Without this, every token from
-// the LLM renders as its own bubble — unreadable for any reply
-// longer than a word or two.
-function CT_coalesceMessages(messages) {
-  const out = [];
-  let buffer = null;
-  const flushBuffer = () => {
-    // Skip text-only buffers whose content is whitespace-only. LLMs
-    // commonly emit one or more empty/zero-delta assistant_token rows
-    // alongside a tool_call (the protocol uses the token row as the
-    // carrier for an empty `content` field when the model's reply is
-    // *all* tool calls). Without this guard, every tool call produces
-    // a blank assistant_message bubble above it — exactly what the
-    // operator reported as "two empty agent messages".
-    if (buffer && buffer.text.trim().length > 0) {
-      out.push(buffer);
-    }
-    buffer = null;
-  };
-  for (const m of messages) {
-    if (m.kind === "assistant_token") {
-      const delta = typeof m.delta === "string" ? m.delta : "";
-      if (!buffer) {
-        buffer = {
-          kind: "assistant_message",
-          text: delta,
-          startSeq: m.seq,
-          endSeq: m.seq,
-        };
-      } else {
-        buffer.text += delta;
-        buffer.endSeq = m.seq;
-      }
-      continue;
-    }
-    flushBuffer();
-    out.push(m);
-  }
-  flushBuffer();
-  return out;
-}
+//
+// Assistant-token coalescing moved to
+// ui/components/chat/use-transcript.js (window.chatCoalesce) as part
+// of Task B2 — it's now shared with <Conversation> and unit-testable
+// on its own.
 
 // Subtle "Thinking…" placeholder shown after the user sends a frame
 // but before the first assistant_token / tool_call / done row lands.
@@ -1659,7 +954,7 @@ function Message({ m }) {
   }
 
   // Coalesced agent reply (the streaming tokens collapsed into one
-  // bubble by CT_coalesceMessages above). Renders as markdown — LLMs
+  // bubble by window.chatCoalesce). Renders as markdown — LLMs
   // routinely emit headings, lists, bold, and code blocks; raw text
   // is borderline unreadable for any non-trivial response.
   if (kind === "assistant_message") {
