+++ ui/index.html
@@ -120,6 +120,8 @@
 <script type="text/babel" src="components/approvals.jsx"></script>
 <script type="text/babel" src="components/channels.jsx"></script>
 <script type="text/babel" src="components/channel_rules.jsx"></script>
+<script type="text/babel" src="components/chat/use-transcript.js"></script>
+<script type="text/babel" src="components/chat/conversation.jsx"></script>
 <script type="text/babel" src="components/chats.jsx"></script>
 <script type="text/babel" src="components/harness_form.jsx"></script>
 <script type="text/babel" src="components/harness_outbound_builder.jsx"></script>
